Fix Virus Motion
Moving the virus up and down when pointed almost directly up
appeared to move the virus sideways when moving up. Moving the
virus in a circle always caused it to slowly swing down and left.

This was due to rounding error- positive decimals floor to 0,
negative decimals floor to 1. So the positive decimal sine and
cosine outputs were silenced, but the negative decimal sine and
cosine outputs always resulted in negative motion (left and down
in this case).